requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==24.2
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.3
